0.0.26 feat: добавлены тесты

+++ src/Kanban/task/Epic.java
@@ -40,7 +40,7 @@ public class Epic extends Task {
     }
 
     public void removeSubtask(Subtask subtask) {
-        subtasks.remove(subtask.getid());
+        subtasks.remove((Integer) subtask.getid());
     }
 
     public void removeSubtasks() {
